requirements/development.txt
# =============================================================================
# Development Requirements
# =============================================================================

-r base.txt

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
pytest-asyncio>=0.23,<1.0
pytest-mock>=3.12,<4.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0
httpx>=0.26,<1.0
responses>=0.24,<1.0

# Code Quality
flake8>=7.0,<8.0
black>=24.1,<25.0
isort>=5.13,<6.0
mypy>=1.8,<2.0
pylint>=3.0,<4.0
pylint-django>=2.5,<3.0

# Type Stubs
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Debugging
django-debug-toolbar>=4.2,<5.0
ipdb>=0.13,<1.0
ipython>=8.20,<9.0

# Documentation
mkdocs>=1.5,<2.0
mkdocs-material>=9.5,<10.0

# Pre-commit
pre-commit>=3.6,<4.0
